# src/swagger_files/QuickCode.Demo.SmsManagerModule.Api-Swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "QuickCode - Demo - SmsManagerModule Api (1.0.0.0)",
    "termsOfService": "https://www.quickcode.net",
    "contact": {
      "name": "Üzeyir Apaydın",
      "url": "https://www.quickcode.net"
    },
    "license": {
      "name": "QuickCode.Net License",
      "url": "https://www.quickcode.net"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/sms-manager-module/otp-messages/count": {
      "get": {
        "tags": [
          "OtpMessages"
        ],
        "operationId": "api-sms-manager-module-otp-messages-count-get",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/sms-manager-module/otp-messages/{id}": {
      "get": {
        "tags": [
          "OtpMessages"
        ],
        "operationId": "api-sms-manager-module-otp-messages--id--get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OtpMessagesDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "OtpMessages"
        ],
        "operationId": "api-sms-manager-module-otp-messages--id--put",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpMessagesDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpMessagesDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/OtpMessagesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "OtpMessages"
        ],
        "operationId": "api-sms-manager-module-otp-messages--id--delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/sms-manager-module/otp-types/{otpTypesId}/otp-messages/{otpMessagesId}": {
      "get": {
        "tags": [
          "OtpTypes"
        ],
        "operationId": "api-sms-manager-module-otp-types--otpTypesId--otp-messages--otpMessagesId--get",
        "parameters": [
          {
            "name": "otpTypesId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "otpMessagesId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OtpTypesOtpMessagesKeyRestResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/sms-manager-module/sms-senders/{smsSendersId}/otp-messages/{otpMessagesId}": {
      "get": {
        "tags": [
          "SmsSenders"
        ],
        "operationId": "api-sms-manager-module-sms-senders--smsSendersId--otp-messages--otpMessagesId--get",
        "parameters": [
          {
            "name": "smsSendersId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "otpMessagesId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SmsSendersOtpMessagesKeyRestResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "OtpMessagesDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "smsSenderId": {
            "type": "integer",
            "format": "int32"
          },
          "otpTypeId": {
            "type": "integer",
            "format": "int32"
          },
          "gsmNumber": {
            "type": "string",
            "nullable": true
          },
          "otpCode": {
            "type": "string",
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "expireSeconds": {
            "type": "integer",
            "format": "int32"
          },
          "messageDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "messageSid": {
            "type": "string",
            "nullable": true
          },
          "dailyCounter": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "OtpTypesOtpMessagesKeyRestResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "smsSenderId": {
            "type": "integer",
            "format": "int32"
          },
          "otpTypeId": {
            "type": "integer",
            "format": "int32"
          },
          "gsmNumber": {
            "type": "string",
            "nullable": true
          },
          "otpCode": {
            "type": "string",
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "expireSeconds": {
            "type": "integer",
            "format": "int32"
          },
          "messageDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "messageSid": {
            "type": "string",
            "nullable": true
          },
          "dailyCounter": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "SmsSendersOtpMessagesKeyRestResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "smsSenderId": {
            "type": "integer",
            "format": "int32"
          },
          "otpTypeId": {
            "type": "integer",
            "format": "int32"
          },
          "gsmNumber": {
            "type": "string",
            "nullable": true
          },
          "otpCode": {
            "type": "string",
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "expireSeconds": {
            "type": "integer",
            "format": "int32"
          },
          "messageDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "messageSid": {
            "type": "string",
            "nullable": true
          },
          "dailyCounter": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "description": "Please enter a valid Api Key",
        "name": "X-Api-Key",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "ApiKey": [ ]
    }
  ]
}